PreparedDataStatus redesign for GetSizeHandler
